Credentials required
I was struggling to understand this part of the documentation. It seems conflicting about whether the credentials prop is required or not. It reads `**credentials** `string` [required]` but that it might not be necessary. I haven't looked into the implementation, but I get an error when not providing `credentials` so I believe that it indeed is required and not optional as described in the text suggested for removal.